requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
celery==5.2.7
certifi==2023.7.22
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.33.0
cron-descriptor==1.4.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-celery-beat==2.5.0
django-celery-results==2.3.1
django-cors-headers==4.2.0
django-import-export==3.2.0
django-timezone-field==5.1
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
kombu==5.3.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
PyJWT==2.7.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
